rename queryGroup to workloadGroup
