+++ plugins/workload-management/src/main/java/org/opensearch/plugin/wlm/rule/attribute_extractor/IndicesExtractor.java
@@ -10,6 +10,7 @@ package org.opensearch.plugin.wlm.rule.attribute_extractor;
 
 import org.opensearch.action.IndicesRequest;
 import org.opensearch.autotagging.Attribute;
+import org.opensearch.rule.RuleAttribute;
 import org.opensearch.rule.attribute_extractor.AttributeExtractor;
 
 import java.util.List;
@@ -30,8 +31,7 @@ public class IndicesExtractor implements AttributeExtractor<String> {
 
     @Override
     public Attribute getAttribute() {
-        // TODO: this will be replaced by WLM defined index_pattern attribute
-        return null;
+        return RuleAttribute.INDEX_PATTERN;
     }
 
     @Override
